requirements.txt
fastapi~=0.115.12
pydantic~=2.11.2
sentence-transformers~=4.0.2
scikit-learn~=1.6.1
requests~=2.26.0
nltk~=3.6.3
newspaper3k~=0.2.8
readability-lxml~=0.8.1
beautifulsoup4~=4.13.3
langdetect~=1.0.9
torch~=2.5.0
transformers~=4.48.0
duckduckgo_search~=2025.4.4
faiss-cpu~=1.10.0
numpy~=2.0.2
ollama~=0.4.7
url-normalize~=2.2.0
pandas~=2.2.3
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
beautifulsoup4==4.13.3
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
deprecation==2.1.0
elastic-transport==8.17.1
elasticsearch==8.14.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
faiss-cpu==1.11.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.2
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hf-xet==1.0.5
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.8.1
numpy==1.24.3
packaging==24.2
pandas==2.0.3
pillow==11.1.0
postgrest==0.19.3
preshed==3.0.9
propcache==0.3.0
psycopg2-binary==2.9.10
pydantic==2.4.2
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.1
PyYAML==6.0.2
rank-bm25==0.2.2
realtime==2.4.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==4.1.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.5
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
